Separate block and inline rendering code
